Broke User Repo into Sleep and Hydration Classes. Created tests for those classes. Updated scripts.
